Introduce and use a Calamares::Branding::setGlobals method.
The method creates a map called "branding" in the global storage, and
inserts an entry for each of the branding strings. This makes the
branding information accessible to the Python modules.

The method is called by CalamaresApplication::initJobQueue.

This is necessary because the Branding class is in libcalamaresui, so
Python modules cannot access it directly.