mount: Fix /run mount.
It turns out that I had messed this up. Issue #173 was addressed anyway
by the followup fix to unpackfs to only exclude the directories that are
actually mounted (which does work), but with this fix, /run is now
handled as intended.

PS: Should the mount job not check the return value of
libcalamares.utils.mount instead of failing silently?