added aptextrapackages module
The aptextrapackages module allows the user to install some packages
from a predefined list. These packages are described in the config file
and have the ability to run pre- and post-install scripts for each
package.

This module assumes apt is installed and will fail the install if it is
not. Thus it is only compatible with Debian and its derivatives.