Change permissions of avatar file after copying.
The permissions are changed to the target user and one
of the default groups (guaranteed to exist on the target
system); chown is invoked synchronously, since Qt does not
support ownership changes.

We also avoid returning errors for any bad conditions, as
this is a less important job in the installation process.