i18n: QT_TRANSLATIONS_DIR hasn't been set since Qt 5.3.
None of the Qt translations are found, nor shipped along with the
Calamares translations; it depends on the translations of Qt
on the live system. The mechanism was broken, anyway, by the
split into qt_<foo>.qm and qtbase_<foo>.qm, and the introduction
for QPT for the translation of standard button texts.

Just drop the special code looking for Qt translations and the
rcc hacks that entails.

FIXES #336