Improve validation of configuration files.
The validation of the configuration files (modules' .conf files,
as well as settings.conf overall) spits out warnings, but does
not do all the validation it can:
 - should print locations more clearly (e.g. if there's a tab
   in welcome.conf)
 - should also print some part of the bad text, if possible
 - if a module cannot be created, that should be treated as
   a fatal error (now the module is silently ignored). This
   is different from a module that can't be found, which
   immediately triggers a fatal failure.

Test plan:
 - introduce a tab into welcome.conf
 - introduce a bad conversion into settings.conf

With just the bad welcome.conf, it should complain clearly, and
refuse to start. With a bad conversion, it seems that there is
a problem but execution can continue, so it should be made clearer
what the problem is. If welcome.conf is buggy, and welcome is
listed to show, this should be a fatal error.