Let pacman do complete system upgardes instead of partial upgardes
While pacman supports partial upgrades, there are no distributions that support
this behaviour. As calamares doesn't support a "upgrade_system" routine we can't
just rely on "pacman -S" being sufficient for an installer, as it might pull in
packages that doesn't exists in the mirror. As "-Sy" might break systems because
of this, we are better off dispatching complete system upgrades instead of
leaving user systems broken.