[unpackfs] Use option -x
The mountpoint directories are not created in the target because they
are marked explicitly as to be excluded using the option --exclude
mountpoint/. They are created later by the module mount (to confirmed).

The option --one-file-system tells rsync to avoid crossing a filesystem
boundary when recursing that is what.

If this option is repeated, rsync omits all mount-point directories from
the copy. Otherwise, it includes an empty directory at each mount-point
it encounters (using the attributes of the mounted directory because
those of the underlying mount-point directory are inaccessible).

This removes the function global_excludes() that creates the list of the
--exclude options and adds the options -x instead (the short option for
--one-file-system).